Add TimeSpan Properties to BaseResponse for Rate Limit Headers (#359)
# Abstract

The current headers return strings such as `12m30s`. If you want to
consume this data, it needs to be parsed into a more usable format. This
PR will add properties to the abstract class BaseResponse that will
parse the header strings into something more useful.

# Changes
- Two new public readonly properties of type TimeSpan have been added to
BaseResponse, `ResetTokensTimespan ` and `ResetTokensTimespan `
- The original properties have not been changed in order to avoid any
breaking changes for downstream projects that may be relying on them.

# Testing Methods
Since regex done poorly can be a massive hindrance in terms of
performance, the regex's performance has been tested using a separate
project. The regex was run through 100k randomly generated timestamps.
The code can be found
[here](https://github.com/RealStillkill/OpenAI-DotNet-RegexMetrics). The
compute time of the regex was negligible.

This change was based on an extension method used in a private project.
It has not produced any problems during its usage in the other project.

---------

Co-authored-by: Stephen Hodgson <rage.against.the.pixel@gmail.com>